feature : BookRecyclerList 리팩토링
