+++ Android/app/src/main/java/com/booktree/BookDetailActivity.java
@@ -15,7 +15,9 @@ public class BookDetailActivity extends AppCompatActivity {
     super.onCreate(savedInstanceState);
     doc = (Documents)getIntent().getSerializableExtra("document");
     setContentView(R.layout.activity_book_detail);
+
     setTopViewGroup();
+
   }
 
   protected void setTopViewGroup() {
